refactor: decompose run.ts into 7 modules (2930 → 320 lines) (#606)
* refactor: extract run-types.ts and run-utils.ts from run.ts (#447)

Step 1-2 of run.ts decomposition:
- run-types.ts: RunOptions, ExecutionContext interfaces + constants
- run-utils.ts: pure utility functions (generateExecutionId, selectMcpConfig,
  detectTaskType, resolveModel, ensureProjectTrusted, getProjectRoot,
  formatDuration, checkClaudeCliAvailable, getClaudeModelAlias)


* refactor: extract execution-log.ts from run.ts (#447)

Step 3: Extract bridge/API helpers, execution logging, cooldown tracking,
and event emission into dedicated module.


* refactor: extract cloud-dispatch.ts from run.ts (#447)

Step 4: Extract self-contained cloud worker dispatch and polling logic.


* refactor: extract execution-engine.ts from run.ts (#447)

Step 5: Extract executeWithClaude, executeWithProvider, and all execution
helpers (worktree management, foreground/background/watch modes, auto-commit,
verification, preflight checks) into dedicated module. -769 lines from run.ts.


* refactor: extract agent-runner.ts and run-modes.ts from run.ts (#447)

Steps 6-7: Extract runAgent() into agent-runner.ts and all squad execution
modes (autopilot, lead, squad loop, post-evaluation) into run-modes.ts.

run.ts: 2930 → 320 lines (89% reduction). Now contains only command
registration, runCommand() routing, and runSquad() dispatch.


---------
